+++ packages/transaction-pay-controller/src/index.ts
@@ -4,6 +4,7 @@ export type {
   TransactionPayControllerActions,
   TransactionPayControllerEvents,
   TransactionPayControllerGetDelegationTransactionAction,
+  TransactionPayControllerGetStrategiesAction,
   TransactionPayControllerGetStateAction,
   TransactionPayControllerGetStrategyAction,
   TransactionPayControllerMessenger,